Code structure refactor
