+++ pages/tabs/others/components/Qualifications.tsx
